Refactor a bit, add a new proc.
I'm communicating with the API, but it's still unclear how to access the
data I'm getting back. Right now it's just pointers. So far research on
"unsafe" and "reflect" aren't helpful.

I think I need to be able to do this to set a callback (i.e. in order to
know which device got the callback and to make sense out of the data the
callback gets), but I can try that next anyway.